[projectm-eval] Fix ns-eel2 shim, adding it as a proper feature
Made the feature default on Windows, as this is the only platform the shim is currently useful on. Users of other platforms can still manually enable the feature if required, e.g. when building with MinGW.

Also added usage information for the optional "Milkdrop" component added by the "ns-eel2" features.